# AERIS/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AERIS Backend",
    "description": "Maryland statewide site screening, scoped parcel investigation, mapped physical constraints, public grid context, and planning context.",
    "version": "0.6.1"
  },
  "paths": {
...
    "/analysis/parcels/build/zone/{zone_id}": {
      "post": {
        "tags": [
          "parcels"
        ],
        "summary": "Build Zone Scope",
        "operationId": "build_zone_scope_analysis_parcels_build_zone__zone_id__post",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          },
          {
            "name": "X-AERIS-Build-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aeris-Build-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildZoneRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/build/bbox": {
      "post": {
        "tags": [
          "parcels"
        ],
        "summary": "Build Bbox Scope",
        "operationId": "build_bbox_scope_analysis_parcels_build_bbox_post",
        "parameters": [
          {
            "name": "X-AERIS-Build-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aeris-Build-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildBBoxRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/jobs/{job_id}": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Build Job",
        "operationId": "parcel_build_job_analysis_parcels_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/status": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Scope Status",
        "operationId": "parcel_scope_status_analysis_parcels_scopes__scope_id__status_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Parcel Scope Status Analysis Parcels Scopes  Scope Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/bundle": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Scope Bundle",
        "operationId": "parcel_scope_bundle_analysis_parcels_scopes__scope_id__bundle_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8000,
              "minimum": 1,
              "default": 8000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/zone/{zone_id}": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcels For Zone",
        "operationId": "parcels_for_zone_analysis_parcels_zone__zone_id__get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8000,
              "minimum": 1,
              "default": 8000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/bbox": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcels For Bbox",
        "operationId": "parcels_for_bbox_analysis_parcels_bbox_get",
        "parameters": [
          {
            "name": "west",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "West"
            }
          },
          {
            "name": "south",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "South"
            }
          },
          {
            "name": "east",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "East"
            }
          },
          {
            "name": "north",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "North"
            }
          },
          {
            "name": "scope_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8000,
              "minimum": 1,
              "default": 8000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/parcels/{parcel_id}": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Detail",
        "operationId": "parcel_detail_analysis_parcels_scopes__scope_id__parcels__parcel_id__get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          },
          {
            "name": "parcel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Parcel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParcelDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/development-envelopes": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Development Envelopes",
        "operationId": "development_envelopes_analysis_parcels_scopes__scope_id__development_envelopes_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/largest-components": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Largest Components",
        "operationId": "largest_components_analysis_parcels_scopes__scope_id__largest_components_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/constraints": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Constraints",
        "operationId": "parcel_constraints_analysis_parcels_scopes__scope_id__constraints_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/parcels/scopes/{scope_id}/grid-evidence": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Parcel Grid Evidence",
        "operationId": "parcel_grid_evidence_analysis_parcels_scopes__scope_id__grid_evidence_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analysis/parcels/scopes/{scope_id}/planning-evidence": {
      "get": {
        "tags": [
          "parcels"
        ],
        "summary": "Planning Evidence",
        "operationId": "planning_evidence_analysis_parcels_scopes__scope_id__planning_evidence_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactStatus": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "ready",
              "missing",
              "failed"
            ],
            "title": "State"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "updated_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Utc"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "state"
        ],
        "title": "ArtifactStatus"
      },
      "BuildBBoxRequest": {
        "properties": {
          "west": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "West"
          },
          "south": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "South"
          },
          "east": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "East"
          },
          "north": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "North"
          },
          "scope_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Name"
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "west",
          "south",
          "east",
          "north"
        ],
        "title": "BuildBBoxRequest"
      },
      "BuildJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "scope_id": {
            "type": "string",
            "title": "Scope Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial_failure",
              "failed"
            ],
            "title": "State"
          },
          "target": {
            "additionalProperties": true,
            "type": "object",
            "title": "Target"
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh"
          },
          "created_at_utc": {
            "type": "string",
            "title": "Created At Utc"
          },
          "updated_at_utc": {
            "type": "string",
            "title": "Updated At Utc"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "artifacts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ArtifactStatus"
            },
            "type": "object",
            "title": "Artifacts"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "job_id",
          "scope_id",
          "state",
          "target",
          "refresh",
          "created_at_utc",
          "updated_at_utc",
          "current_stage",
          "progress",
          "artifacts"
        ],
        "title": "BuildJob"
      },
      "BuildZoneRequest": {
        "properties": {
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "BuildZoneRequest"
      },
...
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Features"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScopeBundle": {
        "properties": {
          "scope_id": {
            "type": "string",
            "title": "Scope Id"
          },
          "artifacts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ArtifactStatus"
            },
            "type": "object",
            "title": "Artifacts"
          },
          "parcels": {
            "$ref": "#/components/schemas/GeoJSONFeatureCollection"
          },
          "envelopes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          },
          "constraints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          },
          "grid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          },
          "planning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "scope_id",
          "artifacts",
          "parcels"
        ],
        "title": "ScopeBundle"
      },
...
    }
  }
}